Replaced the Process Usage Indicator with Led Level Meters for CPU, Mem & I/O Usage

+++ src/rdj/CLUI.java
@@ -753,7 +753,7 @@ public class CLUI implements UI
     {
     }
 
-    @Override public void processProgress(int filesProgress, int fileProgress, long bytesTotalParam, long bytesProcessedParam, long bytesPerMiliSecondParam)
+    @Override public void processProgress(int filesProgress, int fileProgress, long bytesTotalParam, long bytesProcessedParam, double bytesPerMiliSecondParam)
     {
 //        log("filesProgress: " + filesProgress + " fileProgress: " + fileProgress);
     }
